legg til openinghours for flere jobber, unødvendig at de kjører midt på natten uten nytte
