+++ apps/etterlatte-behandling/src/main/kotlin/behandling/jobs/DoedsmeldingReminderJob.kt
@@ -1,12 +1,14 @@
 package no.nav.etterlatte.behandling.jobs
 
+import jobs.OpeningHours
+import jobs.TimerJob
 import no.nav.etterlatte.Context
 import no.nav.etterlatte.Self
 import no.nav.etterlatte.behandling.doedshendelse.DoedshendelseReminderService
 import no.nav.etterlatte.common.DatabaseContext
 import no.nav.etterlatte.jobs.LoggerInfo
 import no.nav.etterlatte.jobs.fixedRateCancellableTimer
-import no.nav.etterlatte.libs.common.TimerJob
+import no.nav.etterlatte.libs.common.tidspunkt.norskKlokke
 import no.nav.etterlatte.sak.SakTilgangDao
 import org.slf4j.LoggerFactory
 import java.time.Duration
@@ -20,6 +22,7 @@ class DoedsmeldingReminderJob(
     private val interval: Duration,
     dataSource: DataSource,
     val sakTilgangDao: SakTilgangDao,
+    private val openingHours: OpeningHours,
 ) : TimerJob {
     private val logger = LoggerFactory.getLogger(this::class.java)
     private val jobbNavn = this::class.simpleName
@@ -36,7 +39,7 @@ class DoedsmeldingReminderJob(
             loggerInfo = LoggerInfo(logger = logger, loggTilSikkerLogg = false),
             period = interval.toMillis(),
         ) {
-            if (erLeader()) {
+            if (erLeader() && openingHours.isOpen(norskKlokke())) {
                 doedshendelseReminderService.setupKontekstAndRun(jobContext)
             }
         }
