+++ apps/etterlatte-behandling/src/main/kotlin/behandling/jobs/SaksbehandlerJob.kt
@@ -1,9 +1,11 @@
 package no.nav.etterlatte.behandling.jobs
 
+import jobs.OpeningHours
+import jobs.TimerJob
 import no.nav.etterlatte.behandling.job.SaksbehandlerJobService
 import no.nav.etterlatte.jobs.LoggerInfo
 import no.nav.etterlatte.jobs.fixedRateCancellableTimer
-import no.nav.etterlatte.libs.common.TimerJob
+import no.nav.etterlatte.libs.common.tidspunkt.norskKlokke
 import org.slf4j.LoggerFactory
 import java.time.Duration
 import java.util.Timer
@@ -13,6 +15,7 @@ class SaksbehandlerJob(
     private val erLeader: () -> Boolean,
     private val initialDelay: Long,
     private val interval: Duration,
+    private val openingHours: OpeningHours,
 ) : TimerJob {
     private val logger = LoggerFactory.getLogger(this::class.java)
     private val jobbNavn = this::class.simpleName
@@ -26,7 +29,7 @@ class SaksbehandlerJob(
             loggerInfo = LoggerInfo(logger = logger, loggTilSikkerLogg = false),
             period = interval.toMillis(),
         ) {
-            if (erLeader()) {
+            if (erLeader() && openingHours.isOpen(norskKlokke())) {
                 saksbehandlerJobService.run()
             }
         }
