+++ apps/etterlatte-behandling/src/main/kotlin/oppgave/FristGaarUtJobb.kt
@@ -1,10 +1,10 @@
 package no.nav.etterlatte.oppgave
 
+import jobs.TimerJob
 import no.nav.etterlatte.funksjonsbrytere.FeatureToggle
 import no.nav.etterlatte.funksjonsbrytere.FeatureToggleService
 import no.nav.etterlatte.jobs.LoggerInfo
 import no.nav.etterlatte.jobs.fixedRateCancellableTimer
-import no.nav.etterlatte.libs.common.TimerJob
 import no.nav.etterlatte.libs.common.oppgave.OppgaveKilde
 import no.nav.etterlatte.libs.common.oppgave.OppgaveType
 import no.nav.etterlatte.libs.common.oppgave.Status
