+++ apps/etterlatte-tidshendelser/src/main/kotlin/no/nav/etterlatte/tidshendelser/JobbPoller.kt
@@ -1,17 +1,17 @@
 package no.nav.etterlatte.tidshendelser
 
+import jobs.OpeningHours
+import jobs.TimerJob
 import no.nav.etterlatte.jobs.LoggerInfo
 import no.nav.etterlatte.jobs.fixedRateCancellableTimer
-import no.nav.etterlatte.libs.common.TimerJob
+import no.nav.etterlatte.libs.common.tidspunkt.norskKlokke
 import org.slf4j.LoggerFactory
-import java.time.Clock
 import java.time.Duration
 import java.util.Timer
 
 class JobbPollerTask(
     private val initialDelaySeconds: Long,
     private val periode: Duration,
-    private val klokke: Clock,
     private val openingHours: OpeningHours,
     private val jobbPoller: JobbPoller,
 ) : TimerJob {
@@ -26,7 +26,7 @@ class JobbPollerTask(
             period = periode.toMillis(),
             loggerInfo = LoggerInfo(logger = logger),
         ) {
-            if (openingHours.isOpen(klokke)) {
+            if (openingHours.isOpen(norskKlokke())) {
                 jobbPoller.poll()
             }
         }
@@ -59,20 +59,5 @@ class JobbPoller(
                 hendelseDao.oppdaterJobbstatusFerdig(jobbRefreshed)
             }
         }
-    }
-}
-
-data class OpeningHours(val start: Int, val slutt: Int) {
-    companion object {
-        fun of(openingHours: String): OpeningHours {
-            openingHours.split("-").toList()
-                .also { require(it.size == 2) }
-                .let { return OpeningHours(it[0].toInt(), it[1].toInt()) }
-        }
-    }
-
-    fun isOpen(klokke: Clock): Boolean {
-        val time = klokke.instant().atZone(klokke.zone).hour
-        return time in start until slutt
     }
 }
