+++ libs/etterlatte-ktor/src/main/kotlin/initialisering/EmbeddedServerConfig.kt
@@ -11,8 +11,8 @@ import io.ktor.server.engine.connector
 import io.ktor.server.engine.embeddedServer
 import io.ktor.server.routing.Route
 import io.ktor.server.routing.routing
+import jobs.TimerJob
 import no.nav.etterlatte.addShutdownHook
-import no.nav.etterlatte.libs.common.TimerJob
 import no.nav.etterlatte.libs.common.logging.sikkerlogger
 import no.nav.etterlatte.libs.ktor.healthApi
 import no.nav.etterlatte.libs.ktor.ktor.shutdownPolicyEmbeddedServer
